coprv2: renamed module to coprocessor_v2
