+++ components/test_raftstore/src/server.rs
@@ -38,7 +38,7 @@ use raftstore::store::{
 use raftstore::Result;
 use security::SecurityManager;
 use tikv::coprocessor;
-use tikv::coprv2;
+use tikv::coprocessor_v2;
 use tikv::import::{ImportSSTService, SSTImporter};
 use tikv::read_pool::ReadPool;
 use tikv::server::gc_worker::GcWorker;
@@ -308,7 +308,7 @@ impl Simulator for ServerCluster {
             concurrency_manager.clone(),
             PerfLevel::EnableCount,
         );
-        let coprv2 = coprv2::Endpoint::new();
+        let coprv2 = coprocessor_v2::Endpoint::new();
         let mut server = None;
         // Create Debug service.
         let debug_thread_pool = Arc::new(
