+++ src/lib.rs
@@ -54,7 +54,7 @@ extern crate test;
 
 pub mod config;
 pub mod coprocessor;
-pub mod coprv2;
+pub mod coprocessor_v2;
 pub mod import;
 pub mod read_pool;
 pub mod server;
